Refactor macOS Run to share implementation between app and Xcode
